MapObjects now can be broken.

+++ core/src/ru/icarumbas/bagel/Utils/B2dWorld/WorldContactListener.kt
@@ -1,17 +1,20 @@
 package ru.icarumbas.bagel.Utils.B2dWorld
 
+import com.badlogic.gdx.audio.Sound
 import com.badlogic.gdx.physics.box2d.*
 import ru.icarumbas.*
+import ru.icarumbas.bagel.Characters.mapObjects.Breakable
 import ru.icarumbas.bagel.Characters.mapObjects.Chest
 import ru.icarumbas.bagel.Characters.mapObjects.PortalDoor
 import ru.icarumbas.bagel.Characters.mapObjects.Spike
-import ru.icarumbas.bagel.Characters.mapObjects.SpikeTrap
 import ru.icarumbas.bagel.Screens.GameScreen
 import kotlin.experimental.or
 
 class WorldContactListener(val gameScreen: GameScreen) : ContactListener {
 
     val deleteList = ArrayList<Body>()
+    var touchedOpeningItem = false
+
 
     override fun postSolve(contact: Contact, impulse: ContactImpulse) {
     }
@@ -36,9 +39,21 @@ class WorldContactListener(val gameScreen: GameScreen) : ContactListener {
                 contact.isEnabled = false
 
                 if (deleteList.isEmpty()) {
+                    gameScreen.assetManager["Sounds/coinpickup.wav", Sound::class.java].play()
+
                     deleteList.add(otherBody)
-                    gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach { if (it is Chest) it.coins.remove(otherBody) }
+                    gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                        if (it is Chest && it.coins.contains(otherBody)) it.coins.remove(otherBody) else
+                        if (it is Breakable && it.coins.contains(otherBody)) it.coins.remove(otherBody)
+                    }
                     gameScreen.player.money += 1
+                }
+            }
+
+            PLAYER_BIT or BREAKABLE_BIT -> {
+                contact.isEnabled = false
+                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                    if (it is Breakable && it.body == otherBody) it.canBeBroken = true
                 }
             }
 
@@ -49,17 +64,27 @@ class WorldContactListener(val gameScreen: GameScreen) : ContactListener {
             }
 
             PLAYER_BIT or CHEST_BIT -> {
-                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach { if (it is Chest && it.body == otherBody) it.isOpened = true }
+                touchedOpeningItem = true
+                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                    if (it is Chest && it.body == otherBody) it.isOpened = true
+                }
                 contact.isEnabled = false
             }
 
             PLAYER_BIT or SPIKE_BIT -> {
-                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach { if (it is Spike && it.body == otherBody) it.isTouched = true }
+                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                    if (it is Spike && it.body == otherBody) it.isTouched = true
+                }
                 contact.isEnabled = false
             }
 
             PLAYER_BIT or PORTAL_DOOR_BIT -> {
-                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach { if (it is PortalDoor) it.opened = true }
+                touchedOpeningItem = true
+                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                    if (it is PortalDoor) {
+                        it.isOpened = true
+                    }
+                }
                 contact.isEnabled = false
             }
         }
@@ -67,11 +92,14 @@ class WorldContactListener(val gameScreen: GameScreen) : ContactListener {
     }
 
     fun deleteBodies(){
-        deleteList.forEach {
-            gameScreen.world.destroyBody(it)
-        }
+        if (deleteList.isNotEmpty()) {
 
-        deleteList.clear()
+            deleteList.forEach {
+                gameScreen.world.destroyBody(it)
+                println("Destroyed body $it.")
+            }
+            deleteList.clear()
+        }
     }
 
     override fun beginContact(contact: Contact) {
@@ -92,9 +120,22 @@ class WorldContactListener(val gameScreen: GameScreen) : ContactListener {
 
         when (fixA.filterData.categoryBits or fixB.filterData.categoryBits) {
 
+            PLAYER_BIT or BREAKABLE_BIT -> {
+                gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach {
+                    if (it is Breakable && it.body == otherBody) it.canBeBroken = false
+                }
+            }
+
             PLAYER_BIT or SPIKE_BIT -> {
                 gameScreen.rooms[gameScreen.currentMap].mapObjects.forEach { if (it is Spike && it.body == otherBody) it.isTouched = false }
+            }
 
+            PLAYER_BIT or CHEST_BIT -> {
+                touchedOpeningItem = false
+            }
+
+            PLAYER_BIT or PORTAL_DOOR_BIT -> {
+                touchedOpeningItem = false
             }
         }
     }
